Call it a baton again.
Back in the day Damian called the structure we use to pass data to and
from the worker thread with a baton but later changed it. I'd like to
bring it back and separate out the field used for writing tags into a
child structure. I think this will set us up nicely for adding a third
structure for reading preview images.